Optimize water rendering, refactor water and terrain views
[#174489917]
[#174236032]
