+++ src/components/view-3d/water-fragment.glsl
@@ -2,4 +2,5 @@ varying vec4 vColor;
 
 void main() {
 	gl_FragColor = vColor;
+	if (gl_FragColor.a < 0.01) discard;
 }
